Move repository creation methods to the API
Update tests and documentation.

fixes #550
