+++ curator/api/utils.py
@@ -5,6 +5,7 @@ import re
 import sys
 import logging
 import json
+import click
 logger = logging.getLogger(__name__)
 
 def get_alias(client, alias):
@@ -298,6 +299,158 @@ def create_snapshot_body(indices, ignore_unavailable=False,
     else:
         body["indices"] = to_csv(indices)
     return body
+
+def delete_callback(ctx, param, value):
+    if not value:
+        ctx.abort()
+
+def create_repo_body(repo_type=None,
+                     compress=True, chunk_size=None,
+                     max_restore_bytes_per_sec=None,
+                     max_snapshot_bytes_per_sec=None,
+                     location=None,
+                     bucket=None, region=None, base_path=None, access_key=None,
+                     secret_key=None, **kwargs):
+    """
+    Build the 'body' portion for use in creating a repository.
+
+    :arg repo_type: The type of repository (presently only `fs` and `s3`)
+    :arg compress: Turn on compression of the snapshot files. Compression is
+        applied only to metadata files (index mapping and settings). Data files
+        are not compressed. (Default: `True`)
+    :arg chunk_size: The chunk size can be specified in bytes or by using size
+        value notation, i.e. 1g, 10m, 5k. Defaults to `null` (unlimited chunk
+        size).
+    :arg max_restore_bytes_per_sec: Throttles per node restore rate. Defaults
+        to ``20mb`` per second.
+    :arg max_snapshot_bytes_per_sec: Throttles per node snapshot rate. Defaults
+        to ``20mb`` per second.
+    :arg location: Location of the snapshots. Required.
+    :arg bucket: `S3 only.` The name of the bucket to be used for snapshots.
+        Required.
+    :arg region: `S3 only.` The region where bucket is located. Defaults to
+        `US Standard`
+    :arg base_path: `S3 only.` Specifies the path within bucket to repository
+        data. Defaults to value of ``repositories.s3.base_path`` or to root
+        directory if not set.
+    :arg access_key: `S3 only.` The access key to use for authentication.
+        Defaults to value of ``cloud.aws.access_key``.
+    :arg secret_key: `S3 only.` The secret key to use for authentication.
+        Defaults to value of ``cloud.aws.secret_key``.
+
+    :returns: A dictionary suitable for creating a repository from the provided
+        arguments.
+    :rtype: dict
+    """
+    # This shouldn't happen, but just in case...
+    if not repo_type:
+        click.echo(click.style('Missing required parameter --repo_type', fg='red', bold=True))
+        sys.exit(1)
+
+    argdict = locals()
+    body = {}
+    body['type'] = argdict['repo_type']
+    body['settings'] = {}
+    settings = []
+    maybes   = [
+                'compress', 'chunk_size',
+                'max_restore_bytes_per_sec', 'max_snapshot_bytes_per_sec'
+               ]
+    s3       = ['bucket', 'region', 'base_path', 'access_key', 'secret_key']
+
+    settings += [i for i in maybes if argdict[i]]
+    # Type 'fs'
+    if argdict['repo_type'] == 'fs':
+        settings.append('location')
+    # Type 's3'
+    if argdict['repo_type'] == 's3':
+        settings += [i for i in s3 if argdict[i]]
+    for k in settings:
+        body['settings'][k] = argdict[k]
+    return body
+
+def create_repository(client, **kwargs):
+    """
+    Create repository with repository and body settings
+
+    :arg client:  The Elasticsearch client connection
+
+    :arg repo_type: The type of repository (presently only `fs` and `s3`)
+    :arg compress: Turn on compression of the snapshot files. Compression is
+        applied only to metadata files (index mapping and settings). Data files
+        are not compressed. (Default: `True`)
+    :arg chunk_size: The chunk size can be specified in bytes or by using size
+        value notation, i.e. 1g, 10m, 5k. Defaults to `null` (unlimited chunk
+        size).
+    :arg max_restore_bytes_per_sec: Throttles per node restore rate. Defaults
+        to ``20mb`` per second.
+    :arg max_snapshot_bytes_per_sec: Throttles per node snapshot rate. Defaults
+        to ``20mb`` per second.
+    :arg location: Location of the snapshots. Required.
+    :arg bucket: `S3 only.` The name of the bucket to be used for snapshots.
+        Required.
+    :arg region: `S3 only.` The region where bucket is located. Defaults to
+        `US Standard`
+    :arg base_path: `S3 only.` Specifies the path within bucket to repository
+        data. Defaults to value of ``repositories.s3.base_path`` or to root
+        directory if not set.
+    :arg access_key: `S3 only.` The access key to use for authentication.
+        Defaults to value of ``cloud.aws.access_key``.
+    :arg secret_key: `S3 only.` The secret key to use for authentication.
+        Defaults to value of ``cloud.aws.secret_key``.
+
+    :returns: A boolean value indicating success or failure.
+    :rtype: boolean
+    """
+    if not 'repository' in kwargs:
+        click.echo(click.style('Missing required parameter --repository', fg='red', bold=True))
+        sys.exit(1)
+    else:
+        repository = kwargs['repository']
+
+    try:
+        body = create_repo_body(**kwargs)
+        logger.info("Checking if repository {0} already exists...".format(repository))
+        result = get_repository(client, repository=repository)
+        logger.debug("Result = {0}".format(result))
+        if not result:
+            logger.info("Repository {0} not in Elasticsearch. Continuing...".format(repository))
+            client.snapshot.create_repository(repository=repository, body=body)
+        elif result is not None and repository not in result:
+            logger.info("Repository {0} not in Elasticsearch. Continuing...".format(repository))
+            client.snapshot.create_repository(repository=repository, body=body)
+        else:
+            logger.error("Unable to create repository {0}.  A repository with that name already exists.".format(repository))
+            sys.exit(1)
+    except elasticsearch.TransportError as e:
+        logger.error("Unable to create repository {0}.  Response Code: {1}.  Error: {2}. Check curator and elasticsearch logs for more information.".format(repository, e.status_code, e.error))
+        return False
+    logger.info("Repository {0} creation initiated...".format(repository))
+    return True
+
+def verify_repository(client, repository=None):
+    """
+    Verify the existence of a repository
+
+    Args:
+        client:  The Elasticsearch client connection
+
+    Kwargs:
+        repository: The Elasticsearch snapshot repository to use
+
+    Returns:
+        A boolean value indicating success or failure.
+    """
+    if not repository:
+        click.echo(click.style('Missing required parameter --repository', fg='red', bold=True))
+        sys.exit(1)
+    test_result = get_repository(client, repository)
+    if repository in test_result:
+        logger.info("Repository {0} creation validated.".format(repository))
+        return True
+    else:
+        logger.error("Repository {0} failed validation...".format(repository))
+        return False
 
 def prune_kibana(indices):
     """Remove any index named `.kibana`, `kibana-int`, or `.marvel-kibana`
